V3.89 (#487)
* Removing td

* Streamlining regression tests to cut runtime in half

* Adding a UTF8 console so we can display special characters in the Windows console

* Porting decimal conversion and mathlib to dd_cascade

* WIP: porting the math libraries to the cascade-based Priest number systems

* adding generators for high precision constants

* Moving the to_string/parse infrastructure to floatcascade so it can be reused (#486)


* bug fixes in to_digits()

* adding to_string/parse to triple- and quad-double

* completing to_string/parse consolidation with regression tests

* Restructuring the code with better naming

* Fixing the date on the session logs

* fun with inverse factorials: parse and div consistently differ by 1ULP

* WIP: code hygiene td and qd cascade

* lining up mathlib across dd/td/qd cascades

* WIP: copy stubbed math functions to cascade number systems

* WIP: phase 2: port exp/log/pow/hyperbolic to cascades

* WIP: phase 3: sqrt and trigonmetry functions

* adding mathlibs to triple- and quad-double cascades and fixing precision issues

* WIP: adding sqrt regression tests to cascades

* fixing cube root for td and qd cascades

* replacing Karp-trick sqrt with real Newton-Raphson iterations

* Updating change/session logs

---------

+++ include/sw/math/stub/abs.hpp
@@ -12,7 +12,7 @@ namespace sw { namespace universal {
 template<typename Scalar,
          typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 Scalar abs(Scalar v) {
-	return std::abs(v);
+	return std::abs(double(v));
 }
 
 }} // namespace sw::universal
