+++ include/sw/math/stub/error_and_gamma.hpp
@@ -12,14 +12,14 @@ namespace sw { namespace universal {
 template<typename Scalar,
 	typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 Scalar erf(Scalar x) {
-	return std::erf(x);
+	return std::erf(double(x));
 }
 
 // Compute the complementary error function: 1 - erf(x)
 template<typename Scalar,
 	typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 Scalar erfc(Scalar x) {
-	return std::erfc(x);
+	return std::erfc(double(x));
 }
 
 }} // namespace sw::universal
