+++ include/sw/math/stub/hypot.hpp
@@ -44,7 +44,7 @@ namespace sw { namespace universal {
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar hypot(Scalar x, Scalar y) {
-		return std::hypot(x, y);
+		return std::hypot(double(x), double(y));
 	}
 
 	// hypotf/hypotd/hypotl ?
